+++ webcomponents/lazy-img/src/components/interfaces/custom-load.ts
@@ -1,0 +1,5 @@
+export interface DeckDeckGoCustomLoad {
+  imgElement: HTMLImageElement;
+  imgSrc?: string;
+  imgSrcSet?: string;
+}